requirements.txt
# Core dependencies from original project
baml-py==0.211.2
pydantic==2.12.3
python-dotenv==1.1.1

# DeepSeek OCR dependencies
torch==2.6.0
torchvision
transformers==4.46.3
tokenizers==0.20.3
einops
addict
easydict
safetensors==0.6.2

# Flash attention for DeepSeek OCR (optional but recommended)
# flash-attn==2.7.3  # Uncomment if you have CUDA and want to build from source

# Image processing
pillow==11.3.0
opencv-python==4.12.0.88

# Utilities
requests==2.32.5
tqdm==4.67.1

# Note: No OpenAI API key needed for this branch!
# This branch uses:
# - DeepSeek OCR (via Hugging Face transformers, runs locally)
# - gpt-oss-20b (via Ollama, runs locally)